requirements.txt
accelerate==1.8.1
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
appdirs==1.4.4
APScheduler==3.11.0
argcomplete==3.6.2
asgiref==3.8.1
attr==0.3.1
attrs==25.3.0
azure-core==1.34.0
azure-storage-blob==12.25.1
black==25.1.0
bleach==5.0.1
boto==2.49.0
boto3==1.38.42
botocore==1.38.42
cachetools==5.5.2
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
cryptography==45.0.4
datamodel-code-generator==0.26.1
datasets==3.6.0
defusedxml==0.7.1
dill==0.3.8
distro==1.9.0
Django==5.1.11
django-annoying==0.10.6
django-cors-headers==4.7.0
django-csp==3.7
django-debug-toolbar==3.2.1
django-environ==0.10.0
django-extensions==3.2.3
django-filter==24.3
django-migration-linter==5.2.0
django-model-utils==4.1.1
django-ranged-fileresponse==0.1.2
django-rq==2.10.3
django-storages==1.12.3
django-user-agents==0.4.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
dnspython==2.7.0
drf-dynamic-fields==0.3.0
drf-flex-fields==0.9.5
drf-generators==0.3.0
email_validator==2.2.0
expiringdict==1.2.2
Faker==37.4.0
fastapi==0.115.13
filelock==3.18.0
frozenlist==1.7.0
fsspec==2025.3.0
genson==1.3.0
google-api-core==2.25.1
google-auth==2.40.3
google-cloud-appengine-logging==1.6.2
google-cloud-audit-log==0.3.2
google-cloud-core==2.4.3
google-cloud-logging==3.12.1
google-cloud-storage==2.19.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
grpc-google-iam-v1==0.14.2
grpcio==1.73.0
grpcio-status==1.73.0
h11==0.16.0
hf-xet==1.1.5
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.33.0
humansignal-drf-yasg==1.21.10.post1
idna==3.10
ijson==3.4.0
importlib_metadata==8.7.0
inflect==5.6.2
inflection==0.5.1
isodate==0.7.2
isort==5.13.2
Jinja2==3.1.6
jiter==0.10.0
jmespath==1.0.1
joblib==1.5.1
jsf==0.11.2
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
label-studio==1.19.0
label-studio-sdk==1.0.16
launchdarkly-server-sdk==8.2.1
lockfile==0.12.2
lxml==5.4.0
lxml_html_clean==0.4.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mpmath==1.3.0
multidict==6.5.0
multiprocess==0.70.16
mypy_extensions==1.1.0
networkx==3.5
nltk==3.9.1
numpy==1.26.4
openai==1.91.0
opencv-python==4.11.0.86
opentelemetry-api==1.34.1
ordered-set==4.0.2
packaging==25.0
pandas==2.3.0
pathspec==0.12.1
pillow==11.2.1
platformdirs==4.3.8
propcache==0.3.2
proto-plus==1.26.1
protobuf==6.31.1
psutil==7.0.0
psycopg2-binary==2.9.10
pyarrow==20.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pyboxen==1.3.0
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
PyJWT==2.10.1
pyRFC3339==2.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-json-logger==2.0.4
pytz==2022.7.1
PyYAML==6.0.2
redis==5.2.1
referencing==0.36.2
regex==2024.11.6
requests==2.32.4
requests-file==2.1.0
requests-mock==1.12.1
rich==14.0.0
rpds-py==0.25.1
rq==1.16.2
rsa==4.9.1
rstr==3.2.2
rules==3.4
s3transfer==0.13.0
safetensors==0.5.3
scikit-learn==1.7.0
scipy==1.16.0
semver==3.0.4
sentencepiece==0.2.0
sentry-sdk==2.30.0
setuptools==80.9.0
six==1.17.0
smart-open==7.1.0
sniffio==1.3.1
sqlparse==0.5.3
starlette==0.46.2
sympy==1.13.1
threadpoolctl==3.6.0
tiktoken==0.9.0
tldextract==5.3.0
tokenizers==0.21.1
toml==0.10.2
torch==2.6.0
tqdm==4.67.1
transformers==4.52.4
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
tzlocal==5.3.1
ua-parser==1.0.1
ua-parser-builtins==0.18.0.post1
ujson==5.10.0
uritemplate==4.2.0
urllib3==1.26.20
user-agents==2.2.0
uvicorn==0.34.3
webencodings==0.5.1
wheel==0.40.0
wrapt==1.17.2
xmljson==0.2.1
xxhash==3.5.0
yarl==1.20.1
zipp==3.23.0
